+++ MoreShipUpgrades/Misc/LategameConfiguration.cs
@@ -34,6 +34,7 @@ namespace MoreShipUpgrades.Misc
     public class LategameConfiguration : SyncedConfig2<LategameConfiguration>
     {
         #region Enabled
+        [field: SyncedEntryField] public SyncedEntry<bool> CARBON_KNEEJOINTS_ENABLED {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> LIFE_INSURANCE_ENABLED {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> RUBBER_BOOTS_ENABLED { get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> OXYGEN_CANISTERS_ENABLED {  get; set; }
@@ -79,6 +80,7 @@ namespace MoreShipUpgrades.Misc
         #endregion
 
         #region Individual
+        [field: SyncedEntryField] public SyncedEntry<bool> CARBON_KNEEJOINTS_INDIVIDUAL {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> RUBBER_BOOTS_INDIVIDUAL {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> OXYGEN_CANISTERS_INDIVIDUAL { get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> SLEIGHT_OF_HAND_INDIVIDUAL {  get; set; }
@@ -109,6 +111,7 @@ namespace MoreShipUpgrades.Misc
         #endregion
 
         #region Initial Prices
+        [field: SyncedEntryField] public SyncedEntry<int> CARBON_KNEEJOINTS_PRICE {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> LIFE_INSURANCE_PRICE {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> RUBBER_BOOTS_PRICE {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> OXYGEN_CANISTERS_PRICE { get; set; }
@@ -158,6 +161,10 @@ namespace MoreShipUpgrades.Misc
         #endregion
 
         #region Attributes
+        [field: SyncedEntryField] public SyncedEntry<string> CARBON_KNEEJOINTS_PRICES {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<string> CARBON_KNEEJOINTS_OVERRIDE_NAME {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<int> CARBON_KNEEJOINTS_INITIAL_CROUCH_DEBUFF_DECREASE {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<int> CARBON_KNEEJOINTS_INCREMENTAL_CROUCH_DEBUFF_DECREASE { get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> LIFE_INSURANCE_PRICES {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> LIFE_INSURANCE_OVERRIDE_NAME {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> LIFE_INSURANCE_INITIAL_COST_PERCENTAGE_DECREASE {  get; set; }
@@ -449,6 +456,7 @@ namespace MoreShipUpgrades.Misc
         [field: SyncedEntryField] public SyncedEntry<string> ITEM_PROGRESSION_BLACKLISTED_ITEMS { get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> ITEM_PROGRESSION_APPARATICE_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> ITEM_PROGRESSION_NO_PURCHASE_UPGRADES {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<string> CARBON_KNEEJOINTS_ITEM_PROGRESSION_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> LIFE_INSURANCE_ITEM_PROGRESSION_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> RUBBER_BOOTS_ITEM_PROGRESSION_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> OXYGEN_CANISTERS_ITEM_PROGRESSION_ITEMS {  get; set; }
@@ -547,6 +555,7 @@ namespace MoreShipUpgrades.Misc
 
             topSection = LguConstants.OVERRIDE_NAMES_SECTION;
             OVERRIDE_UPGRADE_NAMES              = cfg.BindSyncedEntry(topSection, LguConstants.OVERRIDE_NAMES_ENABLED_KEY, LguConstants.OVERRIDE_NAMES_ENABLED_DEFAULT, LguConstants.OVERRIDE_NAMES_ENABLED_DESCRIPTION);
+            CARBON_KNEEJOINTS_OVERRIDE_NAME     = cfg.BindSyncedEntry(topSection, LguConstants.CARBON_KNEEJOINTS_OVERRIDE_NAME_KEY, CarbonKneejoints.UPGRADE_NAME);
             LIFE_INSURANCE_OVERRIDE_NAME        = cfg.BindSyncedEntry(topSection, LguConstants.LIFE_INSURANCE_OVERRIDE_NAME_KEY, LifeInsurance.UPGRADE_NAME);
             RUBBER_BOOTS_OVERRIDE_NAME          = cfg.BindSyncedEntry(topSection, LguConstants.RUBBER_BOOTS_OVERRIDE_NAME_KEY, RubberBoots.UPGRADE_NAME);
             OXYGEN_CANISTERS_OVERRIDE_NAME      = cfg.BindSyncedEntry(topSection, LguConstants.OXYGEN_CANISTERS_OVERRIDE_NAME_KEY, OxygenCanisters.UPGRADE_NAME);
@@ -639,6 +648,19 @@ namespace MoreShipUpgrades.Misc
             #endregion
 
             #region Upgrades
+
+            #region Carbon Kneejoints
+
+            topSection = CarbonKneejoints.UPGRADE_NAME;
+            CARBON_KNEEJOINTS_ENABLED = cfg.BindSyncedEntry(topSection, LguConstants.CARBON_KNEEJOINTS_ENABLED_KEY, LguConstants.CARBON_KNEEJOINTS_ENABLED_DEFAULT, LguConstants.CARBON_KNEEJOINTS_ENABLED_DESCRIPTION);
+            CARBON_KNEEJOINTS_INDIVIDUAL = cfg.BindSyncedEntry(topSection, BaseUpgrade.INDIVIDUAL_SECTION, BaseUpgrade.INDIVIDUAL_DEFAULT, BaseUpgrade.INDIVIDUAL_DESCRIPTION);
+            CARBON_KNEEJOINTS_PRICE = cfg.BindSyncedEntry(topSection, LguConstants.CARBON_KNEEJOINTS_PRICE_KEY, LguConstants.CARBON_KNEEJOINTS_PRICE_DEFAULT);
+            CARBON_KNEEJOINTS_PRICES = cfg.BindSyncedEntry(topSection, BaseUpgrade.PRICES_SECTION, CarbonKneejoints.DEFAULT_PRICES, BaseUpgrade.PRICES_DESCRIPTION);
+            CARBON_KNEEJOINTS_INITIAL_CROUCH_DEBUFF_DECREASE = cfg.BindSyncedEntry(topSection, LguConstants.CARBON_KNEEJOINTS_INITIAL_CROUCH_DEBUFF_DECREASE_KEY, LguConstants.CARBON_KNEEJOINTS_INITIAL_CROUCH_DEBUFF_DECREASE_DEFAULT, LguConstants.CARBON_KNEEJOINTS_INITIAL_CROUCH_DEBUFF_DECREASE_DESCRIPTION);
+            CARBON_KNEEJOINTS_INCREMENTAL_CROUCH_DEBUFF_DECREASE = cfg.BindSyncedEntry(topSection, LguConstants.CARBON_KNEEJOINTS_INCREMENTAL_CROUCH_DEBUFF_DECREASE_KEY, LguConstants.CARBON_KNEEJOINTS_INCREMENTAL_CROUCH_DEBUFF_DECREASE_DEFAULT, LguConstants.CARBON_KNEEJOINTS_INCREMENTAL_CROUCH_DEBUFF_DECREASE_DESCRIPTION);
+            CARBON_KNEEJOINTS_ITEM_PROGRESSION_ITEMS = cfg.BindSyncedEntry(topSection, LguConstants.ITEM_PROGRESSION_ITEMS_KEY, LguConstants.ITEM_PROGRESSION_ITEMS_DEFAULT, LguConstants.ITEM_PROGRESSION_ITEMS_DESCRIPTION);
+
+            #endregion
 
             #region Life Insurance